+++ CHANGELOG.md
@@ -14,6 +14,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added a form validation against the DTO in the platform management of the admin control panel
 - Added a form validation against the DTO in the tag management of the admin control panel
 
+### Added
+
+- Added support for data decimation in the line chart component
+
 ### Changed
 
 - Set the performance column of the holdings table to stick at the end